Move functions to non-deprecated library
Closes #2111

The intention had been to deprecate only the import, not the members,
but the semantics of a deprecated library is that all the members are
deprecated. Move the definitions to a non-deprecated library under
`src/` for conventional package privacy, and deprecate the instead of
the entire library `export`.
